+++ wow_login_messages/src/logon/version_2/realm_type.rs
@@ -54,7 +54,7 @@ impl TryFrom<u8> for RealmType {
             1 => Ok(Self::PlayerVsPlayer),
             6 => Ok(Self::Roleplaying),
             8 => Ok(Self::RoleplayingPlayerVsPlayer),
-            v => Err(crate::errors::EnumError::new("RealmType", v as u32),)
+            v => Err(crate::errors::EnumError::new("RealmType", v as u64),)
         }
     }
 }
