+++ wow_login_messages/src/logon/version_3/security_flag.rs
@@ -44,7 +44,7 @@ impl TryFrom<u8> for SecurityFlag {
         match value {
             0 => Ok(Self::None),
             1 => Ok(Self::Pin),
-            v => Err(crate::errors::EnumError::new("SecurityFlag", v as u32),)
+            v => Err(crate::errors::EnumError::new("SecurityFlag", v as u64),)
         }
     }
 }
