+++ wow_world_base/src/inner/shared/class_vanilla_tbc.rs
@@ -80,7 +80,7 @@ impl TryFrom<u8> for Class {
             8 => Ok(Self::Mage),
             9 => Ok(Self::Warlock),
             11 => Ok(Self::Druid),
-            v => Err(crate::errors::EnumError::new("Class", v as u32),)
+            v => Err(crate::errors::EnumError::new("Class", v as u64),)
         }
     }
 }
