+++ wow_world_base/src/inner/tbc/item_class_and_sub_class.rs
@@ -559,7 +559,7 @@ impl TryFrom<u64> for ItemClassAndSubClass {
             12884901903 => Ok(Self::Holiday),
             17179869199 => Ok(Self::MiscellaneousOther),
             21474836495 => Ok(Self::Mount),
-            v => Err(crate::errors::EnumError::new("ItemClassAndSubClass", v as u32),)
+            v => Err(crate::errors::EnumError::new("ItemClassAndSubClass", v),)
         }
     }
 }
