+++ wow_world_base/src/inner/tbc/skill.rs
@@ -715,7 +715,7 @@ impl TryFrom<u16> for Skill {
             767 => Ok(Self::PetRavager),
             768 => Ok(Self::PetSerpent),
             769 => Ok(Self::Internal),
-            v => Err(crate::errors::EnumError::new("Skill", v as u32),)
+            v => Err(crate::errors::EnumError::new("Skill", v as u64),)
         }
     }
 }
