+++ wow_world_messages/src/world/shared/ai_reaction_vanilla_tbc_wrath.rs
@@ -69,7 +69,7 @@ impl TryFrom<u32> for AiReaction {
             2 => Ok(Self::Hostile),
             3 => Ok(Self::Afraid),
             4 => Ok(Self::Destroy),
-            v => Err(crate::errors::EnumError::new("AiReaction", v),)
+            v => Err(crate::errors::EnumError::new("AiReaction", v as u64),)
         }
     }
 }
