+++ wow_world_messages/src/world/shared/arena_faction_tbc_wrath.rs
@@ -44,7 +44,7 @@ impl TryFrom<u8> for ArenaFaction {
         match value {
             0 => Ok(Self::Horde),
             1 => Ok(Self::Alliance),
-            v => Err(crate::errors::EnumError::new("ArenaFaction", v as u32),)
+            v => Err(crate::errors::EnumError::new("ArenaFaction", v as u64),)
         }
     }
 }
