+++ wow_world_messages/src/world/shared/buy_result_vanilla_tbc_wrath.rs
@@ -79,7 +79,7 @@ impl TryFrom<u8> for BuyResult {
             8 => Ok(Self::CantCarryMore),
             11 => Ok(Self::RankRequire),
             12 => Ok(Self::ReputationRequire),
-            v => Err(crate::errors::EnumError::new("BuyResult", v as u32),)
+            v => Err(crate::errors::EnumError::new("BuyResult", v as u64),)
         }
     }
 }
