+++ wow_world_messages/src/world/shared/charter_type_tbc_wrath.rs
@@ -44,7 +44,7 @@ impl TryFrom<u8> for CharterType {
         match value {
             0 => Ok(Self::Guild),
             1 => Ok(Self::Arena),
-            v => Err(crate::errors::EnumError::new("CharterType", v as u32),)
+            v => Err(crate::errors::EnumError::new("CharterType", v as u64),)
         }
     }
 }
