+++ wow_world_messages/src/world/shared/experience_award_type_vanilla_tbc_wrath.rs
@@ -44,7 +44,7 @@ impl TryFrom<u8> for ExperienceAwardType {
         match value {
             0 => Ok(Self::Kill),
             1 => Ok(Self::NonKill),
-            v => Err(crate::errors::EnumError::new("ExperienceAwardType", v as u32),)
+            v => Err(crate::errors::EnumError::new("ExperienceAwardType", v as u64),)
         }
     }
 }
