+++ wow_world_messages/src/world/shared/guild_event_tbc_wrath.rs
@@ -154,7 +154,7 @@ impl TryFrom<u8> for GuildEvent {
             17 => Ok(Self::GuildBankUpdateMoney),
             18 => Ok(Self::GuildBankMoneyWithdrawn),
             19 => Ok(Self::GuildBankTextChanged),
-            v => Err(crate::errors::EnumError::new("GuildEvent", v as u32),)
+            v => Err(crate::errors::EnumError::new("GuildEvent", v as u64),)
         }
     }
 }
